Provide a README file
John Baldwin <jhb@freebsd.org> converted CSRG's SCCS repository holding
the original BSD history to a Subversion repository, which he made
available in the following tarball:

ftp://ftp.freebsd.org/pub/FreeBSD/development/CSRG/csrg_svn.tbz

Except for this commit, the complete history of this Git repository is
created from John's SVN repository, and the README file added by this
commit is based on the one included in the same SVN repository tarball.

Thanks to Kirk McKusick for preserving this history and to John Baldwin
for converting it to Subversion.